+++ src/packets/baseline.rs
@@ -2,19 +2,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! This module provides types and serialisers for each packet type
-//! defined by the NMRA standard.
+//! This module provides types and serialisers for each "baseline"
+//! packet type defined by the NMRA standard.
 
+use super::{Preamble, Result, SerialiseBuffer};
 use crate::Error;
 use bitvec::prelude::*;
-
-/// Convenient Result wrapper
-pub type Result<T> = core::result::Result<T, Error>;
-
-struct Preamble(BitArr!(for 14, in u8, Msb0));
-
-/// Buffer long enough to serialise any common DCC packet into
-pub type SerialiseBuffer = BitArr!(for 43, in u8, Msb0);
 
 impl Default for Preamble {
     fn default() -> Self {