+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+* Refactored `packets` module to have submodules for each packet class
 ### Deprecated
 ### Removed
 ### Fixed